requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
torch>=1.10.0
transformers>=4.18.0
shap>=0.40.0
matplotlib>=3.5.0
seaborn>=0.11.0
optuna>=2.10.0
mlflow>=1.26.0
tqdm>=4.62.0
requests>=2.27.0
emoji>=1.7.0
Pillow>=9.0.0